Split API logic
