+++ src/http3/mod.rs
@@ -1,6 +1,7 @@
-pub mod client;
-pub mod company_rate_limits;
-pub mod error;
-pub mod response;
-pub mod server;
-pub mod user_rate_limits;
+pub(crate) mod client;
+pub(crate) mod company_rate_limits;
+pub(crate) mod error;
+pub(crate) mod handlers;
+pub(crate) mod response;
+pub(crate) mod server;
+pub(crate) mod user_rate_limits;
